Use Karma in Interactor tests (for now) (#37)
## Motivation

We want to use our tests when we have them. With the current setup, we have tests for Interactors, but they use `@bigtest/cli` to run tests. That package is going to change completely once we move the new daemon into this project. I didn't want to try to figure out how to get the current CLI running on CI to get tests running. 

## Approach

Stop using `@bigtest/cli` to run Interactor tests and switch to using `karma` with HeadlessChrome in GitHub Actions.